Completing the photo update section

+++ app/Http/Controllers/Image/ImageController.php
@@ -10,31 +10,42 @@ use Illuminate\Http\Request;
 class ImageController extends Controller
 {
 
-    public function create_image(Request $request) {
-
+    public function create_image(Request $request)
+    {
         $image = Image::create([
             'image' => $request->image->getClientOriginalName(),
         ]);
-        $image ->addMediaFromRequest('image')->toMediaCollection();
+        $image->addMediaFromRequest('image')->toMediaCollection();
         return response()->json([
             'status' => true,
             'message' => 'image has been added'
         ]);
     }
-    public function update_image(Request $request , $id) {
+
+    public function update_image(Request $request, $id)
+    {
         $image = Image::find($id);
-        $image->clearMediaCollection('default');
-        $image->addMediaFromRequest('image')->toMediaCollection();
-        $image->update([
+        $image->clearMediaCollection();
+        $image->delete();
+        $image->create([
             'image' => $request->image->getClientOriginalName(),
         ]);
+        $image->addMediaFromRequest('image')->toMediaCollection();
+
         return response()->json([
             'status' => true,
             'message' => 'image has been updated'
         ]);
     }
 
-    public function destroy($id) {
+    public function destroy($id)
+    {
+        if (Image::find($id) == null) {
+            return response()->json([
+                'status' => false,
+                'message' => 'There is no such photo to delete'
+            ]);
+        }
         $image = Image::find($id);
         $image->clearMediaCollection('default');
         $image->delete();
@@ -42,6 +53,15 @@ class ImageController extends Controller
         return response()->json([
             'status' => true,
             'message' => 'image has been deleted'
+        ]);
+    }
+
+    public function show()
+    {
+        $images = Image::all();
+        return response()->json([
+            'status' => true,
+            'images' => $images,
         ]);
     }
